+++ cmp-worker-ios/src/iosMain/kotlin/io/github/mobilebytelabs/worker/ios/IosWorkManagerInit.kt
@@ -4,7 +4,7 @@ import io.github.mobilebytelabs.worker.ExperimentalWorkerApi
 import io.github.mobilebytelabs.worker.PlatformWorkManager
 
 @ExperimentalWorkerApi
-fun initIosWorkManager(workerFactory: IosWorkerFactory) {
-    val manager = IosWorkManager(workerFactory)
+fun initIosWorkManager(workerFactory: IosWorkerFactory, config: IosWorkManagerConfig = IosWorkManagerConfig.DEFAULT) {
+    val manager = IosWorkManager(workerFactory, config)
     PlatformWorkManager.configure(manager)
 }